// apps/server/src/features/requirement/enums/requirement-type.enum.ts
export enum RequirementType {
  SYSTEM_REQUIREMENT = "SYSTEM_REQUIREMENT",
  EVENT_SYSTEM_REQUIREMENT = "EVENT_SYSTEM_REQUIREMENT",
  ACTOR_REQUIREMENT = "ACTOR_REQUIREMENT",
  SYSTEM_ACTOR_COMMUNICATION_REQUIREMENT = "SYSTEM_ACTOR_COMMUNICATION_REQUIREMENT",
  CONDITIONAL_REQUIREMENT = "CONDITIONAL_REQUIREMENT",
  RECURSIVE_REQUIREMENT = "RECURSIVE_REQUIREMENT",
  USE_CASE_REFERENCE_REQUIREMENT = "USE_CASE_REFERENCE_REQUIREMENT",
  LOGICAL_GROUP_REQUIREMENT = "LOGICAL_GROUP_REQUIREMENT",
  CONDITIONAL_GROUP_REQUIREMENT = "CONDITIONAL_GROUP_REQUIREMENT",
  SIMULTANEOUS_REQUIREMENT = "SIMULTANEOUS_REQUIREMENT",
  EXCEPTIONAL_REQUIREMENT = "EXCEPTIONAL_REQUIREMENT",
}